Feat : 댓글 삭제 기능 구현

+++ src/api/comments.js
@@ -19,9 +19,13 @@ export const getComments = async (id) => {
 
 //댓글 삭제
 export const deleteComments = async (id) => {
-  console.log(id);
+  const confirmed = window.confirm("정말로 삭제하시겠습니까?");
+  if (!confirmed) {
+    return; // 삭제 취소
+  }
+  // console.log(id);
   const response = await axios.delete(
-    `${process.env.REACT_APP_SERVER_URL}/detailpage/${id}`
+    `${process.env.REACT_APP_SERVER_URL}/comments/${id}`
   );
   return response.data;
 };
